+++ app/Http/Controllers/PermissionController.php
@@ -26,9 +26,13 @@ class PermissionController extends BaseController
      * @return void
      */
     public function index(){
-        $this -> setPageData('Permisssion', 'Permisssion', 'fas fa-th-list');
-        $data = $this->service->index();
-        return view('permission.index', compact('data'));
+        if(permission('permission-access')){
+            $this -> setPageData('Permisssion', 'Permisssion', 'fas fa-th-list');
+            $data = $this->service->index();
+            return view('permission.index', compact('data'));
+        }else{
+            return $this->unauthorizedAccessBlocked();
+        }
     }
 
     /**
@@ -38,12 +42,14 @@ class PermissionController extends BaseController
      * @return void
      */
     public function getDataTableData(Request $request){
-        if($request -> ajax()){
-            $output = $this->service->getDataTableData($request);
-        }else{
-            $output = ['status'=>'error','message'=>'Unauthorize access blocked'];
+        if(permission('permission-access')){
+            if($request -> ajax()){
+                $output = $this->service->getDataTableData($request);
+            }else{
+                $output = ['status'=>'error','message'=>'Unauthorize access blocked'];
+            }
+            return response()->json($output);
         }
-        return response()->json($output);
     }
 
     /**
@@ -55,11 +61,15 @@ class PermissionController extends BaseController
 
     public function store(PermissionRequest $request){
         if($request->ajax()){
-            $result = $this->service->store($request);
-            if($result){
-                return $this->responseJson($status='success',$message='Data has been saved successfull',$data=null,$response_code=204);
+            if(permission('permission-add')){
+                $result = $this->service->store($request);
+                if($result){
+                    return $this->responseJson($status='success',$message='Data has been saved successfull',$data=null,$response_code=204);
+                }else{
+                    return $this->responseJson($status='error',$message='Data can not save',$data=null,$response_code=204);
+                }
             }else{
-                return $this->responseJson($status='error',$message='Data can not save',$data=null,$response_code=204);
+                return $this->responseJson($status='error',$message='Unauthorized access blocked',$data=null,$response_code=401);
             }
         }else{
             return $this->responseJson($status='error',$message=null,$data=null,$response_code=401);
@@ -74,11 +84,15 @@ class PermissionController extends BaseController
      */
     public function edit(Request $request){
         if($request->ajax()){
-            $data = $this->service->edit($request);
-            if($data -> count()){
-                return $this -> responseJson($status='success',$message=null,$data=$data,$response_code=201);
+            if(permission('permission-edit')){
+                $data = $this->service->edit($request);
+                if($data -> count()){
+                    return $this -> responseJson($status='success',$message=null,$data=$data,$response_code=201);
+                }else{
+                    return $this->responseJson($status='error',$message='No Data Found',$data=null,$response_code=204);
+                }
             }else{
-                return $this->responseJson($status='error',$message='No Data Found',$data=null,$response_code=204);
+                return $this->responseJson($status='error',$message='Unauthorized access blocked',$data=null,$response_code=401);
             }
         }else{
             return $this->responseJson($status='error',$message=null,$data=null,$response_code=401);
@@ -94,11 +108,15 @@ class PermissionController extends BaseController
 
     public function update(PermissionUpdateRequest $request){
         if($request->ajax()){
-            $result = $this->service->update($request);
-            if($result){
-                return $this->responseJson($status='success',$message='Data has been updated successfull',$data=null,$response_code=204);
+            if(permission('permission-edit')){
+                $result = $this->service->update($request);
+                if($result){
+                    return $this->responseJson($status='success',$message='Data has been updated successfull',$data=null,$response_code=204);
+                }else{
+                    return $this->responseJson($status='error',$message='Data can not update',$data=null,$response_code=204);
+                }
             }else{
-                return $this->responseJson($status='error',$message='Data can not update',$data=null,$response_code=204);
+                return $this->responseJson($status='error',$message='Unauthorized access blocked',$data=null,$response_code=401);
             }
         }else{
             return $this->responseJson($status='error',$message=null,$data=null,$response_code=401);
@@ -113,11 +131,15 @@ class PermissionController extends BaseController
      */
     public function delete(Request $request){
         if($request->ajax()){
-            $result = $this->service->delete($request);
-            if($result){
-                return $this->responseJson($status='success',$message="Data has been deleted successfull",$data=null,$response_code=200);
+            if(permission('permission-delete')){
+                $result = $this->service->delete($request);
+                if($result){
+                    return $this->responseJson($status='success',$message="Data has been deleted successfull",$data=null,$response_code=200);
+                }else{
+                    return $this->responseJson($status='error',$message='Data can not delete',$data=null,$response_code=204);
+                }
             }else{
-                return $this->responseJson($status='error',$message='Data can not delete',$data=null,$response_code=204);
+                return $this->responseJson($status='error',$message='Unauthorized access blocked',$data=null,$response_code=401);
             }
         }else{
             return $this->responseJson($status='error',$message=null,$data=null,$response_code=401);
@@ -126,11 +148,15 @@ class PermissionController extends BaseController
 
     public function bulkDelete(Request $request){
         if($request->ajax()){
-            $result = $this->service->bulkDelete($request);
-            if($result){
-                return $this->responseJson($status='success',$message="Data has been deleted successfull",$data=null,$response_code=200);
+            if(permission('permission-bulk-delete')){
+                $result = $this->service->bulkDelete($request);
+                if($result){
+                    return $this->responseJson($status='success',$message="Data has been deleted successfull",$data=null,$response_code=200);
+                }else{
+                    return $this->responseJson($status='error',$message='Data can not delete',$data=null,$response_code=204);
+                }
             }else{
-                return $this->responseJson($status='error',$message='Data can not delete',$data=null,$response_code=204);
+                return $this->responseJson($status='error',$message='Unauthorized access blocked',$data=null,$response_code=401);
             }
         }else{
             return $this->responseJson($status='error',$message=null,$data=null,$response_code=401);
