+++ app/Http/Controllers/RoleController.php
@@ -25,8 +25,12 @@ class RoleController extends BaseController
      * @return void
      */
     public function index(){
-        $this -> setPageData('Role', 'Role', 'fas fa-th-list');
-        return view('role.index');
+        if(permission('role-access')){
+            $this -> setPageData('Role', 'Role', 'fas fa-th-list');
+            return view('role.index');
+        }else{
+            return $this->unauthorizedAccessBlocked();
+        }
     }
 
     /**
@@ -36,18 +40,24 @@ class RoleController extends BaseController
      * @return void
      */
     public function getDataTableData(Request $request){
-        if($request -> ajax()){
-            $output = $this->service->getDataTableData($request);
-        }else{
-            $output = ['status'=>'error','message'=>'Unauthorize access blocked'];
+        if(permission('role-access')){
+            if($request -> ajax()){
+                $output = $this->service->getDataTableData($request);
+            }else{
+                $output = ['status'=>'error','message'=>'Unauthorize access blocked'];
+            }
+            return response()->json($output);
         }
-        return response()->json($output);
     }
 
     public function create(){
-        $this -> setPageData('Add Role', 'Create Role', 'fas fa-th-list');
-        $data = $this->service->permissionModuleList();
-        return view('role.create', compact('data'));
+        if(permission('role-add')){
+            $this -> setPageData('Add Role', 'Create Role', 'fas fa-th-list');
+            $data = $this->service->permissionModuleList();
+            return view('role.create', compact('data'));
+        }else{
+            $this->unauthorizedAccessBlocked();
+        }
     }
 
     /**
@@ -59,20 +69,23 @@ class RoleController extends BaseController
 
     public function storeOrUpdate(RoleRequest $request){
         if($request->ajax()){
-            $result = $this->service->storeOrUpdate($request);
-            if($result){
-                if($request -> update_id){
-                    return $this->responseJson($status='success',$message='Data has been updated successfull',$data=null,$response_code=204);
+            if(permission('role-add') || permission('role-add')){
+                $result = $this->service->storeOrUpdate($request);
+                if($result){
+                    if($request -> update_id){
+                        return $this->responseJson($status='success',$message='Data has been updated successfull',$data=null,$response_code=204);
+                    }else{
+                        return $this->responseJson($status='success',$message='Data has been saved successfull',$data=null,$response_code=204);
+                    }
                 }else{
-                    return $this->responseJson($status='success',$message='Data has been saved successfull',$data=null,$response_code=204);
+                    if($request->update_id){
+                        return $this->responseJson($status='error',$message='Data can not update',$data=null,$response_code=204);
+                    }else{
+                        return $this->responseJson($status='error',$message='Data can not save',$data=null,$response_code=204);
+                    }
                 }
-
             }else{
-                if($request->update_id){
-                    return $this->responseJson($status='error',$message='Data can not update',$data=null,$response_code=204);
-                }else{
-                    return $this->responseJson($status='error',$message='Data can not save',$data=null,$response_code=204);
-                }
+                return $this->responseJson($status='error',$message='Unauthorized access blocked',$data=null,$response_code=401);
             }
         }else{
             return $this->responseJson($status='error',$message=null,$data=null,$response_code=401);
@@ -86,10 +99,14 @@ class RoleController extends BaseController
      * @return void
      */
     public function edit(int $id){
-        $this -> setPageData('Edit Role', 'Edit Role', 'fas fa-th-list');
-        $data = $this->service->permissionModuleList();
-        $permission_data = $this->service->edit($id);
-        return view('role.edit', compact('data', 'permission_data'));
+        if(permission('role-edit')){
+            $this -> setPageData('Edit Role', 'Edit Role', 'fas fa-th-list');
+            $data = $this->service->permissionModuleList();
+            $permission_data = $this->service->edit($id);
+            return view('role.edit', compact('data', 'permission_data'));
+        }else{
+            $this->unauthorizedAccessBlocked();
+        }
     }
 
     /**
@@ -99,10 +116,14 @@ class RoleController extends BaseController
      * @return void
      */
     public function show(int $id){
-        $this -> setPageData('Role Details', 'Role Details', 'fas fa-th-list');
-        $data = $this->service->permissionModuleList();
-        $permission_data = $this->service->edit($id);
-        return view('role.view', compact('data', 'permission_data'));
+        if(permission('role-show')){
+            $this -> setPageData('Role Details', 'Role Details', 'fas fa-th-list');
+            $data = $this->service->permissionModuleList();
+            $permission_data = $this->service->edit($id);
+            return view('role.view', compact('data', 'permission_data'));
+        }else{
+            $this->unauthorizedAccessBlocked();
+        }
     }
 
     /**
@@ -113,13 +134,17 @@ class RoleController extends BaseController
      */
     public function delete(Request $request){
         if($request->ajax()){
-            $result = $this->service->delete($request);
-            if($result == 1){
-                return $this->responseJson($status='success',$message="Data has been deleted successfull",$data=null,$response_code=200);
-            }elseif($result == 2){
-                return $this->responseJson($status='error',$message="Data can not delet becouse it's releted with many users",$data=null,$response_code=200);
+            if(permission('role-delete')){
+                $result = $this->service->delete($request);
+                if($result == 1){
+                    return $this->responseJson($status='success',$message="Data has been deleted successfull",$data=null,$response_code=200);
+                }elseif($result == 2){
+                    return $this->responseJson($status='error',$message="Data can not delet becouse it's releted with many users",$data=null,$response_code=200);
+                }else{
+                    return $this->responseJson($status='error',$message='Data can not delete',$data=null,$response_code=204);
+                }
             }else{
-                return $this->responseJson($status='error',$message='Data can not delete',$data=null,$response_code=204);
+                return $this->responseJson($status='error',$message='Unauthorized access blocked',$data=null,$response_code=401);
             }
         }else{
             return $this->responseJson($status='error',$message=null,$data=null,$response_code=401);
@@ -128,13 +153,17 @@ class RoleController extends BaseController
 
     public function bulkDelete(Request $request){
         if($request->ajax()){
-            $result = $this->service->bulkDelete($request);
-            if($result['status'] == 1){
-                return $this->responseJson($status='success',$message = !empty($result['message']) ? $result['message'] : "Data has been deleted successfull",$data=null,$response_code=200);
-            }elseif($result == 2){
-                return $this->responseJson($status='error', $message = !empty($result['message']) ? $result['message'] : "Selected data can not delete" , $data=null,$response_code=200);
+            if(permission('role-bulk-delete')){
+                $result = $this->service->bulkDelete($request);
+                if($result['status'] == 1){
+                    return $this->responseJson($status='success',$message = !empty($result['message']) ? $result['message'] : "Data has been deleted successfull",$data=null,$response_code=200);
+                }elseif($result == 2){
+                    return $this->responseJson($status='error', $message = !empty($result['message']) ? $result['message'] : "Selected data can not delete" , $data=null,$response_code=200);
+                }else{
+                    return $this->responseJson($status='error',$message= !empty($result['message']) ? $result['message'] : 'Selected data can not delete',$data=null,$response_code=204);
+                }
             }else{
-                return $this->responseJson($status='error',$message= !empty($result['message']) ? $result['message'] : 'Selected data can not delete',$data=null,$response_code=204);
+                return $this->responseJson($status='error',$message='Unauthorized access blocked',$data=null,$response_code=401);
             }
         }else{
             return $this->responseJson($status='error',$message=null,$data=null,$response_code=401);
